chore: update button types
